Location implemented, RudderGame implemented
-Before this commit, code expected String vertex to access graph nodes,
this effected Game implementation badly. It would be lots of parsing
and converting to String. To avoid it, created Location class as an
interface between Node's and Piece's locations.
-RudderGameLocation, in RudderGame there is a special segment,
Segment.CENTER. In normal conditions, Location's hashCode is
calculating by using segment and level. However, in RudderGame there
is no level for CENTER so level will not be involved to calculation.
For that reason RudderGame has its own Location class.
-RudderGame, as mentioned above, RudderGame class created. It is an
implemented version of Game abstract class. (Still in progress, some
methods still are implementing)
-RudderGameTest class is implemented to test RudderGame.