mostly fixes bugs
-Graph, I'm currently learning Java 8 streams in this project and I did
not know how map works properly. I changed as a proper usage.(Still
not be a good solution, open to any review)
-RudderGame, a few(or lots of) bugging fixed.
In constructor, capturedPiece assigned to a LinkedList.
In initiliazeBoard, the connection between center and segments which
are level 1 is overlooked. Connected with a new Edge.
In initiliazeGame, there was a big problem in here. i.e player 1 was
assigned pieces like, center-1-a2-b3-c4 which we know it is all wrong.
The code changed now, it assings like a1-a2-a3-a4-b1-b2...-d3-d4.
In canMove, first condition is modified. if(currentExist)
In getBetween, this one also worked all wrong. It even could not
determine a2 to b3 is not valid move. After testing codes I saw the
problem and changed the code in a way that works well.