bug fixes, activePlayers implemented.
-RudderGame.capturableMoveExist(..) created. Some of codes in
mandatoryMove() moved in capturableMoveExist(). This needed for the
scenario, p1 captured a piece. If he capture another piece, still p1's
turn. So he can capture twice in a row. To determine this I seperated
this two methods. To more detail check code.

-Game players Set changed to a sequence list.
-Game has a new abstract method that activePlayer(). Games plays
between users in a row. This method gives an answer to question,
"who is turn is it?".
-RudderGame implemented activePlayer(). Only active player can make a
move.
-Bug fixed in RudderGame.determineMoveType().
-capturedPieces convert to Set. There is a recursive call in move().
Sometimes 30 times adds some object to capturedPieces. To avoid it,
using Set.
-capturedPieces did not clear after removed piece on board. It is
cleared now.