Piece are not in Node's anymore.
-Before this, there is a Piece information in each Node. I found that
this is not a good approach. They must be seperated. Only game should
know information about Location->Piece. getAttachedPiece and etc was
not convenient. For this purpose, there is a Map field in Game class.
That maps Location to Piece. The entire code modified with these
changes(if modification is needed).
-If a move is successfull, we add the move into moves Collection.

-Move class has a new field, capture. It represents if move type is
CAPTURE, then one of opponents piece's is captured. The captured
piece puts in new field.
-Game.LEVEL moved to RudderGame class.
-Some classes fields and methods was not accessible outside of
package. These fields modified as public.(Game,RudderGame, etc.)
-Player has a new constructor with parameters.
-Test classes are modified with new changes.