started UI.
-Created App class to starts the game.
-Decided to use JavaFX(first time). RudderGameApp an application class
of JavaFX.
-RudderGameController, this logic layer controls the game with user
interface. Methods' hiearchy is not very well now, it might be well
designed in future commits. Basicly, creating circles and maps the
circles and Locations. Circles are creating and locating dynamicly.
-Circle's have drag and drop event. Players can make a move with
dragging and dropping. If a user can not make a move, highlighting
target location as red, if available as green. However, if there is a
mandatory move, user do not get any responds. He just can't make the
move.
-RudderGame methods are modified. They do same thing just improved
for clean code.
-RudderGameLocation has a new method. segmentDistance(..) it returns
distance between location and callers location.
-Modified test methods.