determineMoveType bug fix
-When a mandatoryMove is detected, the code was adding capturable
Piece to capturedPieces. However, this is a bug. If two mantaryMove
exists(A1-B1 and CENTER-B1) and code offers A1toB1, but, players makes
the other one CENTER-B1. When code offered A1toB1, it adds to
capturablePiece, as you understand player didnt make that move so
the code will remove captured piece from the move A1toB1. For this
purpose before adding a captured pieces, clearing Collection and later
adds.