started impl of Socket programming.
-RudderGameController has a new method that move(). It does the same
thing when player dragging a piece to another. The code pulled in
drag events and put in this method. Drag events call this method now.

-New classes inside of net package. Server is a server that listens
a port. It designed as singleton. After starting to listen if a
client is reached, put the client into a queue. For purpose of
listening clients, created a class that ClientListener. This class
inherits from ServerThread. ServerThread is using for starting and
stopping threads. ClientListener thread will be listen client's
socket to IO process.
-Matchmaking class is matching clients for a game. It creates a
game session to controlling game area in server. This class will be
notify clients. Each game has its own Matchmaking because they have
different rules.
-Most of classes are not done, i'm not sure that is this commit
valid because not done class implementation.