Server-Client imp, Connected network and UI
-Lots of changing in this commit. Project deadline is today. I've been
working on this commit a few days, I tried to move fast on these days.
My last codes, network and connection between network and ui and etc.
can be not well structured.
-Used JSON for transmitting message between Server and Client.
Jackson library responses to do that. I added a annoation to each
class which can be convertable to JSON. Created a class that converts
JSON byte to Message object and object to JSON byte, MessageParser
-GameSession class, this class is simulating games in Server. It
manages moves that coming fromplayers and after a validation(if
move is valid), sends it to other players on this session.
-Client class contains a Socket which is connected Client's socket.
This a bridge between Client and Server. Each I/O sending doing
in here. Also, this class works on a seperated thread from main app.
-ClientObserver class, using Observer pattern in here. When a
Message comes to Client's socket, some classes needs to notified
with this message. This class takes this response. Project should
be designed for several observers, however, there is one observer
at current time. Might be in future commit.
-RudderGame can play offline and online. I patched here, might be
cause a problem in future.
-RudderGameController, init code moved to a seperate method for
allowing playing offline.
This class is an observer now. When a message comes to it, update()
method invokes.